Working alpha beta

+++ src/board.rs
@@ -197,12 +197,12 @@ impl Board {
         self.halfmove_clock_stack.pop().unwrap()
     }
 
-    pub fn score(&self) -> i32 {
+    pub fn score(&self) -> f32 {
         self.material_value()
     }
 
-    pub fn material_value(&self) -> i32 {
-        i32::from(self.white.material_value()) - i32::from(self.black.material_value())
+    pub fn material_value(&self) -> f32 {
+        f32::from(self.white.material_value()) - f32::from(self.black.material_value())
     }
 }
 
@@ -213,6 +213,6 @@ mod tests {
     #[test]
     fn test_starting_material_value() {
         let board = Board::starting_position();
-        assert_eq!(0, board.material_value());
+        assert_eq!(0., board.material_value());
     }
 }
